Split strategy area -> owner / area
A `BaseStrategy`'s `.area` used to be the owning area. This lead to 
unexpected results when used in offering and trading where the area of 
the related markets is more relevant.

Therefore this concept has now been split into `.owner` - the `Area` 
which "owns" the strategy, and `.area` - the `Area` in which the strategy trades.

This means that now in calls to `market.offer()` and 
`market.accept_offer()` `self.owner.name` must be used instead of
`self.area.name`