GSYE-183: Validate matching_requirements to ensure they exist in the Bid/Offer object
The purpose of this commit is to make sure that a matching_requirement sent by the
myco-sdk actually exists in the original Bid/Offer object of the market without any changes.
This commit also proposes that the 'match_recommendations' function overrides the bid/offer
attribute of the BidOfferMatch object provided by the myco-sdk.