GSYE-15: Apply DRY to single/batch methods
Methods with name `_{action}_aggregator` are mostly same as methods with
name `_{action}_impl`. The only difference is that methods of the first group
don't publish the result to redis channel, instead a caller function will gather
all the responses and sends all responses in one batch. On the other hand, the second
group publishes the result as soon as it's finished executing the job. Thus, functionality
of the groups of %95 same, resulting in DRY being violated. This commit suggests a way to
respect DRY while keeping the both functionalities.